+++ python/idtap_api/classes/automation.py
