+++ python/idtap_api/classes/note_view_phrase.py
