+++ python/idtap_api/classes/trajectory.py
