+++ python/idtap_api/pyproject.toml
@@ -1,3 +1,7 @@
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta:__legacy__"
+
 [project]
 name = "idtap_api"
 version = "0.1.0"
@@ -6,3 +10,7 @@ dependencies = ["requests", "pymongo", "google-auth-oauthlib"]
 
 [project.optional-dependencies]
 test = ["pytest", "responses"]
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["idtap_api*"]
