+++ python/idtap_api/tests/articulation_test.py
@@ -4,7 +4,7 @@ import pytest
 
 sys.path.insert(0, os.path.abspath('.'))
 
-from python.api.classes.articulation import Articulation
+from python.idtap_api.classes.articulation import Articulation
 
 # Tests mirror src/js/tests/articulation.test.ts
