+++ python/idtap_api/tests/auth_test.py
@@ -6,7 +6,7 @@ from pathlib import Path
 sys.path.insert(0, os.path.abspath('.'))
 
 import responses
-from python.api.client import SwaraClient
+from python.idtap_api.client import SwaraClient
 
 BASE = 'https://swara.studio/'
