+++ python/idtap_api/tests/chikari_test.py
@@ -2,8 +2,8 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath('.'))
 
-from python.api.classes.chikari import Chikari
-from python.api.classes.pitch import Pitch
+from python.idtap_api.classes.chikari import Chikari
+from python.idtap_api.classes.pitch import Pitch
 
 # Test mirrors src/js/tests/chikari.test.ts
