+++ python/idtap_api/tests/pitch_test.py
@@ -1,7 +1,7 @@
 import os
 import sys
 sys.path.insert(0, os.path.abspath('.'))
-from python.api.classes.pitch import Pitch
+from python.idtap_api.classes.pitch import Pitch
 import math
 import pytest
