+++ python/idtap_api/tests/section_test.py
@@ -3,10 +3,10 @@ import sys
 
 sys.path.insert(0, os.path.abspath('.'))
 
-from python.api.classes.section import Section
-from python.api.classes.phrase import Phrase
-from python.api.classes.trajectory import Trajectory
-from python.api.classes.pitch import Pitch
+from python.idtap_api.classes.section import Section
+from python.idtap_api.classes.phrase import Phrase
+from python.idtap_api.classes.trajectory import Trajectory
+from python.idtap_api.classes.pitch import Pitch
 
 
 # Tests mirror src/ts/tests/section.test.ts
